requirements.txt
prefect==2.0.0
mlflow==2.0.0
fastapi==0.85.0
uvicorn==0.18.2
pandas
scikit-learn
numpy
PyYAML
pytest
matplotlib